Initial rough port of aqsis elliptical filtering code.
This patch implements elliptical weighted averaging for high quality
anisotropic filtering.  So far it's quite rough around the edges: it
doesn't support lots of things it should, like interpolating between
miplevels and wrap modes.  It's also got a fair bit of code formatting
which is at odds with the OIIO style, and may pull in too many of the
aqsis texture utility headers.