Peroidic and black wrap modes for EWA filtering
Wrap modes are implemented by cutting the filter invocation into
multiple pieces, one for each periodic repetition of the image.  For the
black wrap mode, periodic repetitions with non-zero displacement are
treated as black.  The WrapClamp wrap mode is currently ignored and
assumed to correspond to the periodic case.

There are still some small problems with this: artifacts are apparent
near the horizon of the texttex perspective warp, and in corner where
four periodic repetitions of the the texture join.