Bounding slab optimization for EWA filtering
Diagonally anisotropic filters don't fit well into an axis-aligned
bounding box.  This implements a bounding slab instead so that we can
bound the nonzero areas of the filter much more accurately.  For highly
diagonally anisotropic filter invocations the improvement in speed is
more than a factor of two.

Also improve the way that anisotropic clamping is done to reflect the
superior (less blurry) method used in the default anisotropic lookup
scheme.