Clamp wrap mode for EWA filtering
This is implemented by clamping the filter support onto the border of
the original image, and modifying the filter coefficients so that
they're equal to the integral of the coefficients perpendicular to the
boundary.  All combinations of black, periodic, and clamp modes now seem
to work pretty much correctly, though there's still some oddities near
the horizon of the testtex warp.